requirements.txt
numpy
requests
Pillow
pandas
torch
torchvision
tqdm
scikit-learn
matplotlib
torch-model-archiver